+++ view/frontend/web/js/view/payment/method-renderer/basev2.js
@@ -274,17 +274,33 @@ define(
                 const shipping = quote.shippingAddress();
 
                 const customerId = window.checkoutConfig?.customerData.id || '';
-                const uniqueCustomerId = `${customerId}_${email}_${shop}`;
+                const uniqueCustomerId = !quote.guestEmail && customerId ? `${customerId}_${email}_${shop}` : '';
 
                 if (unzerCustomerId) {
                     this.customer = unzerCustomerId;
                 }
+
+                const mapSalutation = (prefix) => {
+                    if (!prefix) return "unknown";
+
+                    const normalized = prefix.toLowerCase().trim();
+
+                    if (normalized.startsWith('mrs') || normalized.startsWith('ms') || normalized.startsWith('mis')) {
+                        return 'mrs';
+                    }
+                    if (normalized.startsWith('mr')) {
+                        return 'mr';
+                    }
+
+                    return "unknown";
+                };
 
                 const customer = {
                     id: unzerCustomerId || '',
                     customerId: uniqueCustomerId,
                     firstname: billing ? billing.firstname : '',
                     lastname: billing ? billing.lastname : '',
+                    salutation: billing ? mapSalutation(billing.prefix) : 'unknown',
                     email: email,
                     ...(customerData?.dob ? {birthDate: customerData.dob.split('T')[0]} : {}),
                     billingAddress: billing ? {