+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.0.5](https://github.com/unzerdev/magento2/compare/4.0.4..4.0.5)
+### Fixed
+*  Fix customerId assignment for guest customers in UPL payment methods
+*  Fix and improve salutation handling logic
+
 ## [4.0.4](https://github.com/unzerdev/magento2/compare/4.0.3..4.0.4)
 ### Fixed
 *  Resolve issue with customer company information mapping